Testing of Embed Messages
Discord makes use of RichPresence, which we can use to make our general messages look professional and unique. As a result of this, I've been tinkering with simple embed messages (see the help command).

Additionally, the mentors file is where all of the mentor related commands will be stored as functions.