Major Redesign
I was previously working by a heavily outdated version. This new version uses a lot more modern methods found in the latest version of the API. It also showcases pretty much every command type or action that I could want to do, so this acts as a basic template for all work.